app/requirements.txt
boto3
botocore
s3fs
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
et-xmlfile==1.1.0
idna==3.4
jmespath==1.0.1
joblib==1.2.0
kaggle==1.5.13
numpy==1.24.3
opendatasets==0.1.22
openpyxl==3.1.2
pandas==2.0.1
pathlib==1.0.1
pytest==7.3.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
requests==2.30.0
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
streamlit==1.22.0
text-unidecode==1.3
threadpoolctl==3.1.0
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
xlrd==2.0.1
fsspec==2023.5.0